Dynamically resolve dependencies urls for Bootstrap Launcher
The Bootstrap Launcher was designed to download anything from Maven Central.
But since commit 97e67cb has added JDA for Discord the Bootstrap Launcher
was not able to download it because it along with some of its dependencies are only hosted on JCenter.
This commit adds the ability to resolve dependencies from other repos,
 but currently only JCenter and Maven Central are used.
It still tries to get anything from Maven Central and only uses JCenter,
 or any other repo that may be used in the future, as a fallback
 if Maven Central doesn't have a dependency.