Create config dir, if it isn't existent in main framework instead of Bootstrap Launcher
On the first start the default config should be saved, which fails if the folder doesn't exist.
This creates a problem when you try to login to the gui.
The framework can't be loaded because the config with plugins and connectors
don't exist and you won't be able to login, until you create the config directory and restart.

A check for the config dir was in the Bootstrap Launcher, but when
 you run Chat Overflow in IntelliJ that check isn't working.
So I added that check in the main framework.