Fix the support for plugin dependencies inside the plugin dev environment
The PomInclusionPlugin didn't work in the plugin developer environment
because sbt seems to only check the unmanagedBase for auto-plugins,
not unmanagedJars that was used to load the build jar.
This commit moves the jar of the build code in the plugin dev environment
from `bin/chatoverflow-meta-build-*.jar` to `project/lib/chatoverflow-meta-build-*.jar`.
By this the jar is in the unmanagedBase directory and sbt loads
the PomInclusionPlugin and the dependency support for plugins is working
in the plugin dev environment.