Add exact gui jar file to the classpath instead of a wildcard
With the wildcard IntelliJ would only add the gui jar to the classpath after
a reload of the sbt project, because the gui jar does only exist after
an advanced build, but not at the normal import time.

With this commit IntelliJ will try to add the jar on the first reload/import
even tho it currently doesn't exist. That's not a problem, IntelliJ
will just ignore the nonexistent jar and once the advanced build has been
run IntelliJ will add the jar to the classpath without a reload.

Note that a reload is still required after the gui version number has changed!

This commit also drops the scala version suffix in the name of the gui jar
because it doesn't contain any scala code and having the scala suffix
in the name does not make any sense.