Split Bootstrap Updater into a extra project and jar
While the approach to integrate the updater worked on linux (and maybe all *nix systems)
it doesn't work for Windows, because Windows has file locks on running jars
meaning the Updater can't override the old bootstrap launcher with
the new one while it is running.

Now the Updater is the main entry point, which checks for updates.
If found he asks the user, if he wants to install it or not.
After that (updated or not) he starts the Launcher jar,
which then does dependency resolution and starts the actual framework.