First fetch then compile in advanced builds
This was probably a cause of compile problems where the plugins.sbt was missing
and the advanced build then couldn't compile with the error that the framework
can't find any classes of the api. This is because advanced was first compiling, which
requires the api and then fetching the api, which provides it.