Provide fallback dependencies for the framework if no plugins.sbt exists
This allows for compiling when the plugins.sbt hasn't been fetched yet.

This should also fix some IntelliJ errors which could happen if
IntelliJ imported the project while no plugins.sbt exists.
In theses cases IntelliJ wouldn't add the gui and api to the
dependencies of the framework and try to compile/run it without them.
This results in compile errors because the framework requires the api
and the gui wouldn't be included in the classpath which results
in a non working gui.
With the fallback IntelliJ always gets the api and gui dependencies
and the plugins are compiled, packaged and copied by sbt anyway.