Move framework dependency fallback configuration into a separate build definition
Sadly my alphabetical ordering trick didn't work as wanted and on some installations
it was applied after the plugins.sbt which resulted in a situation where no plugins could be built.

By moving it into a separate build definition in the build project we ensure that the
fallback is first applied and can then be overwritten by the plugins.sbt
because first the meta-build settings are applied and then all *.sbt files.